Extract parser core to McsFileParserCore (Node E)
Create McsFileParserCore in McsCore with shared parsing algorithm:
InternalCompileFile, GetMetaDataInfo, DeriveSchemaName,
HashStringToGuid, GetSubAgentComponentId. Move IMcsFileParser interface
from Sync to McsCore. Reduce SyncMcsFileParser and McsFileParser
(Impl.Language.CopilotStudio) to thin adapters that delegate to
McsFileParserCore with consumer-specific error factories
(InvalidOperationException vs UnsupportedBotElementException).
